add custom binding annotation for snapshot action service
When a snapshot create request comes in check if the snapshot count for
the account (orion account) is over quota. If so reject the request. If
the count is below the quota (open slots) then accept creation
request as usual. However, if the count is on the threshold then mark
the oldest snapshot for deprecation and once the new snapshot creation
completes, destroy the old snapshot that was marked for deprecation.

Phase2 tests for snapshot resource

Phase2 tests for snapshot creation orch plugin

Phase2 tests for snapshot core module